chore: add typed helpers for Object.xxxx utils
