+++ packages/utils/src/index.ts
@@ -54,9 +54,8 @@ export * from './splitStringEveryNCharacters';
 export * from './throttler';
 export * from './throwError';
 export * from './topologicalSort';
+export * from './typedObject';
 export * from './typedEventEmitter';
-export * from './typedObjectFromEntries';
-export * from './typedObjectKeys';
 export * from './urlToOnion';
 export * from './zip';
 export * from './removeTrailingSlashes';
