--- packages/utils/src/typedObjectKeys.ts
@@ -1,2 +1,0 @@
-export const typedObjectKeys = <T extends Record<any, any>>(obj: T): Array<keyof T> =>
-    Object.keys(obj) as Array<keyof T>;
