+++ packages/utils/tests/typedObjectValues.type-test.ts
@@ -1,14 +1,14 @@
-import { typedObjectKeys } from '../src/typedObject';
+import { typedObjectValues } from '../src/typedObject';
 
 type AB = { a: 'A'; b: 'B' } | { b: 'BB' };
 
-type ExpectedType = 'a' | 'b';
+type ExpectedType = 'BB' | 'B' | 'A';
 
 let _assertExpectedType: ExpectedType[];
 
-const test1 = typedObjectKeys({ b: 'BB' } satisfies AB);
+const test1 = typedObjectValues({ b: 'BB' } satisfies AB);
 _assertExpectedType = test1;
-const test2 = typedObjectKeys({ a: 'A', b: 'B' } satisfies AB);
+const test2 = typedObjectValues({ a: 'A', b: 'B' } satisfies AB);
 _assertExpectedType = test2;
 
 // eslint-disable-next-line no-self-assign
